requirements.txt
streamlit
google-generativeai
python-pptx
google-cloud-texttospeech
requests
